requirements.txt
yfinance>=0.2.36
requests>=2.31.0
google-genai>=1.0.0
anthropic>=0.40.0
textual>=1.0.0
rich>=13.0.0
